[Feature] Orientacion correcta del personaje y solucionado el Bug que trajo consigo

+++ Assets/Scripts/NoMonoBehavior/Messenger.cs
@@ -14,7 +14,7 @@ public class Messenger
     public bool ApplyForce;
     public bool ApplyInertia;
     public bool Jumping;
-    public short Walking;
-    public short ComboCount;
+    public sbyte Walking;
+    public sbyte ComboCount;
     public AnimationStates RequestedAttack;
 }
